dce-jobs-plugin/requirements.txt
Flask==0.12
dce-plugin-sdk
docker==2.3.0
gunicorn==19.6.0
requests==2.10.0
Flask-RESTful==0.3.5
Flask-Cors==2.1.2
python-etcd==0.4.3
